requirements.txt
MySQL-python
pymssql
impacket
requests
pysnmp
pycrypto
paramiko
python-ldap
pymongo
psycopg2
redis
IPy
THC-Hydra